Move things that require go-jack to pkg/jackutils
So that other things using pkg/common don't have to pull in the dependency
